refactor(ui/tui): move Cline account dialog into @cline/ui/tui
Structural AccountDialogSnapshot/AccountDialogOrganization types keep the
package free of @cline/core; credit formatting moves to
@cline/ui/tui/formatting and the CLI keeps only the account data services
