+++ apps/cli/src/utils/output.ts
@@ -196,19 +196,3 @@ export function writeErr(text: string): void {
 // Formatting helpers
 // =============================================================================
 
-export function formatCreditBalance(value: number, decimalPlaces = 2): string {
-	if (!Number.isFinite(value)) {
-		return "$0.00";
-	}
-	return `$${value.toLocaleString("en-US", {
-		minimumFractionDigits: decimalPlaces,
-		maximumFractionDigits: decimalPlaces,
-	})}`;
-}
-
-export function normalizeCreditBalance(value: number): number {
-	if (!Number.isFinite(value)) {
-		return 0;
-	}
-	return value / 1_000_000;
-}
